third_party/WebKit/LayoutTests/crypto/subtle/cryptokey-interface-is-visible.html

<!DOCTYPE html>
<html>
<head>
<script src="../../resources/js-test.js"></script>
<script src="resources/common.js"></script>
<script src="resources/keys.js"></script>
</head>
<body>
<p id="description"></p>
<div id="console"></div>

<script>
description("The CryptoKey interface should be visible to scripts");

jsTestIsAsync = true;

crypto.subtle.importKey('raw', new Uint8Array(16), "AES-CBC", true, ['encrypt']).then(function(result) {
    key = result;
    shouldBeTrue("key instanceof CryptoKey", true);
}).then(finishJSTest, failAndFinishJSTest);

</script>

</body>
</html>
